+++ src/examples/segway/BlockSegwayController.hpp
@@ -2,7 +2,7 @@
 
 #include <px4_posix.h>
 #include <controllib/blocks.hpp>
-#include <controllib/uorb/blocks.hpp>
+#include "blocks.hpp"
 
 using control::BlockPI;
 using control::BlockP;
